+++ CHANGELOG.md
@@ -1,4 +1,27 @@
 # Release Notes
+## 0.4.1
+
+* add endpoint getUsersByBusinessUnit
+
+## 0.4.0
+
+* endpoint position/getUserPosition/getAdminPosition
+
+## 0.3.9
+
+* add get user by rolename; fixed
+
+## 0.3.8
+
+* add get user by rolename
+
+## 0.3.7
+
+* sticker 404 endpoint issue fixed
+
+## 0.3.6
+
+* deserialized issue fixed
 
 ## 0.4.2